requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.1

scikit-surprise==1.1.4

matplotlib==3.9.0
seaborn==0.13.2

Flask==3.0.3
gunicorn==22.0.0